+++ packages/react-ui/src/app/features/authentication/components/sign-in-form.tsx
@@ -25,7 +25,7 @@ import {
   OpsEdition,
   SignInRequest,
 } from '@openops/shared';
-import { lastVisitedUtils } from '../lib/last-visited';
+import { lastVisitedUtils } from '../../../lib/last-visited';
 import { emailRegex } from '../lib/password-validation-utils';
 
 const SignInSchema = Type.Object({
