+++ packages/react-ui/src/app/lib/authentication-session.ts
@@ -3,7 +3,7 @@ import { jwtDecode } from 'jwt-decode';
 
 import { authenticationApi } from '@/app/lib/authentication-api';
 import { AuthenticationResponse, isNil } from '@openops/shared';
-import { lastVisitedUtils } from '../features/authentication/lib/last-visited';
+import { lastVisitedUtils } from './last-visited';
 import { projectAuth } from './project-auth';
 
 const tokenKey = 'token';
